Function mistralrs_vision::get_resize_image_size

pub fn get_resize_image_size(
    (h, w): (usize, usize),
    (min_len, max_len): (usize, usize),
) -> (usize, usize)

Given the image sizes (h, w) and the minimum and maximum lengths, calculate the image dimensions which will preserve aspect ration while respecing the minimum and maximum lengths.